Making C++ compliant. 'CC=g++ make' for C++23.
 C++ Compatibility Changes vs 405e442
  ┌────────────────────────────────┬───────┬───────────┐
  │            Category            │ Count │  Safety   │
  ├────────────────────────────────┼───────┼───────────┤
  │ void* → typed* casts           │ ~45   │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ malloc/realloc casts           │ ~25   │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ int -> enum casts              │ ~20   │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ enum loop -> int + cast        │ 3     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Case label braces {}           │ 12    │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Struct tags for extern arrays  │ 6     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Macro declaration parens       │ 2     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Designated initializer removal │ 1     │ Low       │
  ├────────────────────────────────┼───────┼───────────┤
  │ Compound literal workaround    │ 1     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ SDL type casts                 │ 3     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Unsigned literal (-1 -> ~0u)   │ 1     │ Safe      │
  ├────────────────────────────────┼───────┼───────────┤
  │ Narrowing cast in braced init  │ 2     │ Safe      │
  └────────────────────────────────┴───────┴───────────┘

+++ src/vga256d.h
@@ -34,7 +34,7 @@ void JE_barBright(SDL_Surface *surface, int a, int b, int c, int d);
 
 static inline void fill_rectangle_wh(SDL_Surface *surface, int x, int y, uint w, uint h, Uint8 color)
 {
-	SDL_Rect rect = { x, y, w, h };
+	SDL_Rect rect = { x, y, (int)w, (int)h };
 	SDL_FillRect(surface, &rect, color);
 }
